requirements.txt
google-generativeai
langchain
langchain-google-genai
yfinance
python-dotenv
requests